scripts/requirements-azure-storage-test.txt
azure-identity>=1.24.0
azure-storage-blob>=12.27.0